+++ tests/test_unit.py
@@ -9,7 +9,6 @@
 # SPDX-License-Identifier: MPL-2.0
 #
 # This file is part of the Antares project.
-from typing import Any
 
 import numpy as np
 import numpy.testing as npt
@@ -46,7 +45,7 @@ def base_cluster_365_days():
     return ThermalCluster(
         unit_count=10,
         nominal_power=100,
-        modulation=np.ones(dtype=float, shape=24),
+        modulation=np.ones(dtype=float, shape=8760),
         fo_law=ProbabilityLaw.UNIFORM,
         fo_volatility=0,
         po_law=ProbabilityLaw.UNIFORM,
@@ -93,7 +92,7 @@ def valid_cluster() -> ThermalCluster:
     return ThermalCluster(
         unit_count=10,
         nominal_power=100,
-        modulation=np.ones(dtype=float, shape=24),
+        modulation=np.ones(dtype=float, shape=8760),
         fo_law=ProbabilityLaw.UNIFORM,
         fo_volatility=0,
         po_law=ProbabilityLaw.UNIFORM,
@@ -169,7 +168,7 @@ def test_forced_outages(rng):
     cluster = ThermalCluster(
         unit_count=10,
         nominal_power=100,
-        modulation=np.ones(dtype=float, shape=24),
+        modulation=np.ones(dtype=float, shape=8760),
         fo_law=ProbabilityLaw.UNIFORM,
         fo_volatility=0,
         po_law=ProbabilityLaw.UNIFORM,
@@ -204,7 +203,7 @@ def test_planned_outages(rng):
     cluster = ThermalCluster(
         unit_count=10,
         nominal_power=100,
-        modulation=np.ones(dtype=float, shape=24),
+        modulation=np.ones(dtype=float, shape=8760),
         fo_law=ProbabilityLaw.UNIFORM,
         fo_volatility=0,
         po_law=ProbabilityLaw.UNIFORM,
@@ -238,7 +237,7 @@ def test_planned_outages_limitation(rng):
     cluster = ThermalCluster(
         unit_count=10,
         nominal_power=100,
-        modulation=np.ones(dtype=float, shape=24),
+        modulation=np.ones(dtype=float, shape=8760),
         fo_law=ProbabilityLaw.UNIFORM,
         fo_volatility=0,
         po_law=ProbabilityLaw.UNIFORM,
@@ -271,7 +270,7 @@ def test_planned_outages_min_limitation(rng):
     cluster = ThermalCluster(
         unit_count=10,
         nominal_power=100,
-        modulation=np.ones(dtype=float, shape=24),
+        modulation=np.ones(dtype=float, shape=8760),
         fo_law=ProbabilityLaw.UNIFORM,
         fo_volatility=0,
         po_law=ProbabilityLaw.UNIFORM,