+++ CHANGELOG.md
@@ -1,12 +1,17 @@
 Changelog
 =========
 
+v0.1.3 (2023-08-05)
+-------------------
+
+* Remove unused imports
+* Use 8760 values for `cluster.modulation` instead of 24.
+
 v0.1.2 (2023-08-02)
 -------------------
 
 * Add `py.typed` file to avoid mypy issue in projects importing the code
 * Fix little typo inside `README.md` file.
-
 
 v0.1.1 (2023-08-02)
 -------------------